requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
# Для асинхронного доступа
asyncpg>=0.27.0,<1.0.0
websockets>=10.0
openai>=1.3.0
# Или для синхронного
# psycopg2-binary>=2.9.5,<3.0.0

# --- База данных ---
SQLAlchemy[asyncio]>=2.0.0 # ORM с поддержкой asyncio
psycopg2-binary>=2.9.5 # Синхронный драйвер для Alembic
alembic>=1.11.0       # Миграции
python-dotenv>=0.20.0 # Для загрузки .env в Alembic
# -------------------

sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
redis>=4.6.0
celery>=5.3.0
flower>=2.0.0 